+++ BeginnerProject/Assets/Scripts/Health.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b1797e62bc8ed8943aaca85e9105fe15
+guid: b6ae10254083f874cac41c3456079f69
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
