+++ BeginnerProject/Assets/Scripts/HealthSystem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b6ae10254083f874cac41c3456079f69
+guid: 25b76a5139482334280bb9265a1fe614
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
