+++ BeginnerProject/Assets/Scripts/PlayerHealth.cs
@@ -3,15 +3,13 @@ using System;
 
 public class PlayerHealth : MonoBehaviour, IDamageable
 {
-    public Health _playerHealth;
-    private  int _damageAmount;
-    private const int _healAmount = 20;
+    private HealthSystem _playerHealth;
     private int _startHealthAmount = 100;
     private int _maxHealthAmount = 100;
     
     private void Awake() 
     {
-        _playerHealth = new Health(_startHealthAmount,_maxHealthAmount);
+        _playerHealth = new HealthSystem(_startHealthAmount,_maxHealthAmount);
     }
 
     private void Heal(int healing)
@@ -30,6 +28,7 @@ public class PlayerHealth : MonoBehaviour, IDamageable
             Dead();
         }
     }
+
     private void Dead()
     {
         throw new NotImplementedException();
